[Test] : addSchedule 테스트코드 추가
Git : #16
