+++ api/eb_fast_api/domain/auth/login/sources/login_router.py
@@ -3,7 +3,7 @@ from eb_fast_api.domain.auth.login.sources import login_feature
 from eb_fast_api.snippets.sources import pwdcrypt
 from eb_fast_api.database.sources.crud import getDB
 from eb_fast_api.domain.auth.login.sources.login_schema import Token
-from eb_fast_api.domain.schema.user_info import UserInfo
+from eb_fast_api.domain.schema.sources.schema import UserInfo
 
 
 router = APIRouter(prefix="/auth/login")
