+++ api/eb_fast_api/domain/auth/login/tests/login_router_test.py
@@ -1,6 +1,6 @@
 from fastapi.testclient import TestClient
 from eb_fast_api.main import app
-from eb_fast_api.domain.schema.user_info import UserInfo
+from eb_fast_api.domain.schema.sources.schema import UserInfo
 from eb_fast_api.database.sources.crud import getDB
 
 
@@ -26,10 +26,8 @@ def test_login_ERROR_invalid_password(loginMockDB):
     testClient = TestClient(app)
 
     # when
-    json = {
-        "email": email,
-        "password": password + "errorString",
-    }
+    userInfo.password += 'errorString'
+    json = userInfo.model_dump(mode='json')
     response = testClient.post("/auth/login", json=json)
 
     # then
@@ -52,10 +50,7 @@ def test_login_SUCCESS(loginMockDB):
     testClient = TestClient(app)
 
     # when
-    json = {
-        "email": email,
-        "password": password,
-    }
+    json = userInfo.model_dump(mode='json')
     response = testClient.post("/auth/login", json=json)
 
     # then
