+++ api/eb_fast_api/domain/schedule/sources/schedule_router.py
@@ -1,14 +1,14 @@
 from fastapi import APIRouter, HTTPException, Depends
 from eb_fast_api.domain.schedule.sources import schedule_feature
 from eb_fast_api.database.sources.crud import getDB
-from eb_fast_api.domain.schema.schedule_info import ScheduleInfo
+from eb_fast_api.domain.schema.sources.schema import ScheduleInfo
 
 
 router = APIRouter(prefix="/schedule")
 
 
 @router.post("/add")
-async def add_schedule(
+async def addSchedule(
     userEmail: str,
     scheduleInfo: ScheduleInfo,
     db=Depends(getDB),
@@ -20,8 +20,9 @@ async def add_schedule(
             db,
         )
     except:
+        # 다른 에러 상황은?
         raise HTTPException(
-            status_code=500,
+            status_code=400,
             detail="서버 스케줄 생성 에러",
         )
